+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Coinbase Pro parser: added "conversion" and "rebate" types.
 - Coinbase parser: added "Reward Income" and "Inflation Reward" transaction types.
 - Coinbase parser: recognise "Coinbase Earn" transactions from the Notes.
+- Conversion tool: added parser for Subscan explorer. ([#305](https://github.com/BittyTax/BittyTax/issues/305))
 ### Removed
 - Conversion tool: removed merge parser for Coinbase/Coinbase Pro.